Исправлены ошибки наставника.

+++ src/main/java/ru/skypro/homework/service/CommentService.java
@@ -2,15 +2,14 @@ package ru.skypro.homework.service;
 
 import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.dto.Comments;
-import ru.skypro.homework.dto.CreateOrUpdateAd;
 import ru.skypro.homework.dto.CreateOrUpdateComment;
 
 public interface CommentService {
-    Comments get(Integer id);
+    Comments get(Long id);
 
-    Comment create(Integer id, CreateOrUpdateComment newComment);
+    Comment create(Long id, CreateOrUpdateComment newComment);
 
-    void delete(Integer adId, Integer commentId);
+    void delete(Long adId, Long commentId);
 
-    Comment update(Integer adId, Integer commentId, CreateOrUpdateComment newComment);
+    Comment update(Long adId, Long commentId, CreateOrUpdateComment newComment);
 }
